shippable: make "go test" exit code visible
This is to avoid repeating this test run:

  https://app.shippable.com/runs/5702d49577ac7a0c0014dce7

The run seems successful, but the "go test" command actually failed.
The problem is that the exit code of "go test" got ignored because we
piped its output to tee. This meant that we couldn't actually see when
the command failed completely -- which is different from the tests
themselves failing.

If "go test" had produced output, any failing tests would have been
reported via the -set-exit-code flag to go-junit-report. However, in
this case "go test" failed to produce any output and go-junit-report
don't report anything when executed on an empty file.