ExecutedCommand: switch ActualOutput to []string
We used [][]byte before and that was messy. Byte slices were
originally chosen since command output can be anything -- not
necessarily UTF-8 encoded text.

However, it turns out[1] that the Go string type is a thin wrapper
over an immutable byte slice. In particular, Go strings can contain
arbitrary bytes. The bytes are UTF-8 decoded when iterated over or
when printed, but otherwise left alone.

This allows us to store the lines of program output as strings, which
in turn simplifies memory management, comparisons, and the tests.

[1]: https://blog.golang.org/strings