main: add --interactive flag
This flag will make Cram prompt you to accept the changes when a command
gives unexpected output. The test file is rewritten inplace.

The user is currently prompted on a per-command basis and it would be
nice to have a way to accept all changes to a test file at once (and to
reject them all too). That will come later.