cram: rename Result struct to ExecutedTest
This struct references the Test struct, mirrowing how ExecutedCommand
references Command. This will eventually allow us to drop the Path field
from Command.

Note that unlike ExecutedCommand, the Test struct is embedded into
ExecutedTest as a value, not a pointer. This is so that the zero value
of ExecutedTest is immediatedly useful.