tests: add ParseTest fuzz test
I was aware that ParseTest couldn't handle invalid test files, but the
go-fuzz framework found the crash in less than a second. We now run the
fuzz test on every build to avoid unexpected regressions in the future.

We only run the fuzz test with Go 1.6 since go-fuzz doesn't support
earlier versions as far as I can tell from various bug reports in their
repo. We still run the normal tests on Go versions 1.2 to 1.5.