ProcessFailures: show diff of output
Instead of showing the expected and actual output as two separate pieces
of text, we now show a unified diff instead.

The diff is not minimized, which means that all common lines are shown.
This is like saying to diff(1) that you want a very large context. This
might be okay for command line output, which is typically short.

Also, assuming the test files are under version control, there is an
easy work-around: use --interactive and run the diff command of your
version control system afterwards. That will give you a diff with the
common lines trimmed.