main: add --jobs flag for parallel testing
We now run cram.Process in parallel from several goroutines. Tests are
often I/O bound which means that we can lower the overall running time
by running multiple tests in parallel.

Some of our test files had to be updated to limit the number of parallel
tests to one so that the output from Cram remains deterministic.

This fixe #20.