MakeBanner: move exit code before random part
Before, the output from the banner command would look something like

    --- CRAM 12345678-1234-abcd-1234-123412345678 --- 42\n

where the "42" part (the exit code) was unpredictable. The banner
command output lines now end with a fixed string instead:

    --- CRAM 42 12345678-1234-abcd-1234-123412345678 ---\n

This allows us to recognize these lines by looking for a suffix instead
of a prefix, which in turn is important to be able to recognize command
output that doesn't end with a newline. Such output will now produce a
line like this:

    command output--- CRAM 42 12345678-1234-abcd-1234-123412345678 ---\n

Since we split on "\n", we can recognize such a line based on its
suffix.