main: return cli.ExitCoder, don't call os.Exit
The call to os.Exit meant that the run function didn't return. Without a
return, the defered os.RemoveAll was not run.

We now use the cli.ExitCoder interface to return a message and an exit
code. The cli library will print the message to stderr (which seems to
be a okay place to put this kind of output) and set the exit code.

Fixes #37.