cram: add support for (no-eol) output lines
The ParseOutput function no longer returns the command output as-is,
instead it now normalizes it to the representation expected to be found
in the test file. This means that " (no-eol)\n" is appended to an
incomplete final output line.

No changes are needed in the rest of the code since it already deals
correctly with the difference.

Changing the output like this means that we cannot distinguish between

  $ echo -n "hello"
  hello (no-eol)

and

  $ echo "hello (no-eol)"
  hello (no-eol)

Both of these tests will pass. That is compatible with the how output is
matched in the ExecutedCommand.failed method: with no syntax for
escaping, it starts by comparing the actual and expected output lines
directly before looking for the special (re) or (glob) suffixes.

Fixes #29.