cram: add Escape and Unescape functions
These functions work similarly to strconv.Quote and
strconv.Unquote (which they use internally), but they add/remove a
"(esc)" suffix to the text. They also don't consider `"` and `\` to be
non-printable characters, which means that they won't be escaped.

This is in preparation for parsing output lines marked with "(esc)".