cram: add support for (esc) lines
We now escape the actual output from the commands as we run them. The
Escape function is a no-op if all characters are printable, but if
something needs escaping, it will do so and add "(esc)". By transforming
the actual output like this, we get correct diffs and patching.

However, we need to update the ExecutedCommand.failed method slightly to
take the escaped lines into account. We need to normalize them since
there are multiple different ways of escaping the same output line.

Fixes #30.