main: add --verbose flag
This flag shows the test file names as they are executed. Unlike a lot
of other test runners, we show the test status as the first character to
keep a nice column alignment:

    . passed.t: 3 commands passed
    F failure.t: 2/5 commands failed
    E error.t:10: an error happened here

This was inspired by Green: https://github.com/CleanCut/green

We will add some color to this later (that is #25 and #26).